requirements.txt
beautifulsoup4==4.11.2
etext==0.0.1
numpy==1.24.0
pandas==1.5.3
plotly==5.10.0
pyparsing==3.0.9
python_dateutil==2.8.2
requests==2.28.1
schedule==1.1.0
scipy==1.10.1
